Kwarg fixes for ruby 3 depreciation warnings
These warnings were appearing on our production logs, and the latest versions of lookbook don't appear to solve them, so I've forked this repo at our current version and added the
patches we need.

Note - upgrading to 1.4+ seemed to completely break our implementation, and I wasn't able to easily get it working. So this is a temporary fix until we can.